Move monitors to dedicated directory
