+++ xtrack/monitors/__init__.py
@@ -1,2 +1,3 @@
 
-from before_loss_monitor import *
+from .particles_monitor import *
+from .before_loss_monitor import *
